# Conflicts:
#	src/main/resources/petriNets/petriflow_schema.xsd

+++ src/main/groovy/com/netgrif/workflow/startup/PostalCodeImporter.groovy
@@ -17,6 +17,9 @@ class PostalCodeImporter extends AbstractOrderedCommandLineRunner {
 
     private static final Logger log = LoggerFactory.getLogger(PostalCodeImporter.class.name)
 
+    @Value("\${postal.codes.import}")
+    Boolean importPostalCode
+
     @Value("\${postal.codes.csv}")
     String postalCodesPath
 
@@ -27,27 +30,31 @@ class PostalCodeImporter extends AbstractOrderedCommandLineRunner {
     private PostalCodeRepository repository
 
     void run(String... strings) {
-        log.info("Importing postal codes from file " + postalCodesPath)
-        def importFile = new ClassPathResource(postalCodesPath).inputStream
+        if (importPostalCode) {
+            log.info("Importing postal codes from file " + postalCodesPath)
+            def importFile = new ClassPathResource(postalCodesPath).inputStream
 
-        def lineCount = 0
-        importFile.readLines().each {
-            lineCount++
+            def lineCount = 0
+            importFile.readLines().each {
+                lineCount++
+            }
+
+            if (repository.count() == lineCount) {
+                log.info("All $lineCount postal codes have been already imported")
+                return
+            }
+
+            repository.deleteAll()
+
+            importFile = new ClassPathResource(postalCodesPath).inputStream
+            def codes = []
+            importFile.splitEachLine(',') { items ->
+                codes << new PostalCode(items[0].replaceAll("\\s", "").trim(), items[1].trim())
+            }
+            service.savePostalCodes(codes)
+            log.info("Postal codes imported")
+        } else {
+            log.info("Skip import postal codes")
         }
-
-        if(repository.count() == lineCount){
-            log.info("All $lineCount postal codes have been already imported")
-            return
-        }
-
-        repository.deleteAll()
-
-        importFile = new ClassPathResource(postalCodesPath).inputStream
-        def codes = []
-        importFile.splitEachLine(',') { items ->
-            codes << new PostalCode(items[0].replaceAll("\\s", "").trim(), items[1].trim())
-        }
-        service.savePostalCodes(codes)
-        log.info("Postal codes imported")
     }
 }
